Consolidate mac_dot into dot (#8)
* Consolidate mac_dot into dot

The majority of functionality and installs are the same between the
hardware. This should make it easier to keep the systems in-check.

* git improvements

add rerere
fix autocorrect

* Document Local Neovim Configs

There must be a better way but this works for now.

* Recipe for LazyVim setup

Create recipe for lazyvim config. I want to remap the keys but telecope
and other plugins do not play well with the simple answer... Once I'm
more familiar I'll try again